reduced redundancy: method calls from profile and timeline cs files
